Separating a line of interest, Easier to read.
Not sure we need the 'safe items' line. further this line doesn't appear to have been updated in a loooong time. 
As I understand it, the 'Safe items' are items that shouldn't be disenchanted. while I've not been able to find some of the existing items that I cherry picked ID's from, I have tested against some items in my bags by adding them. It does stop the disenchanting of listed items.

Now this line functions fine when spaced as such on my end. not getting any errors visibly. 

This line has not been maintained or added to in a while and seems to be a safety net for somewhat random items. 
first item: http://www.wowhead.com/item=89394/amber-spear-of-klaxxivess
BoP item that can be bought again if lost.

TL;DR 
This change tidies it for now.
This line isn't updated, should we maintain/ignore/remove it.