#123 Adding and updating currencies to report.
Have added missing currencies to report
Reordered and commented currencies to make it more legible. Is now in id order. so should be oldest content first. 
Have adjusted some currency caps to match current values. 

--
line 96  Shadowy coins. not able to confirm if actually in use. my rogue has none and most people don't seem to even know it exists. however I'm not seeing a not on Wowhead saying it is not in use. 
Have left it in in case I'm just missing it. can be removed later to tidy things up.

Tooltip example. we might be starting to collect too much info here. 
![tooltip](http://i.imgur.com/PO4JDXM.jpg)

Are we able to get grouping for panel? While in Alphabetical order it still painful to look at.
![panel](http://i.imgur.com/3YYv9Rj.jpg)